+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## v0.7.0 (in development)
 
 - Added a way to disable any minifier.
+- Added a way to detect the default minifier by its options.
+- Added a way to set the default minifier by passing `true` instead of its options.
 
 ## v0.6.0 (2014-11-27)